Extract events mapping  to new EventsPayloadMapper
