+++ src/domain/events/mapper/Models.ts
@@ -1,0 +1,10 @@
+import { Ref } from "@eyeseetea/d2-api";
+
+/**
+ * Models to use by EventsPayloadMapper
+ */
+
+export interface ProgramStageRef {
+    id: string;
+    program: Ref;
+}
