+++ src/domain/events/mapper/__tests__/data/event_program/destination_programs_stages/programStages.json
@@ -1,0 +1,8 @@
+[
+    {
+        "id": "psQUpMGyBEB",
+        "program": {
+            "id": "aaQUpMGyBEB"
+        }
+    }
+]
