+++ src/domain/events/usecases/EventsSyncUseCase.ts
@@ -2,19 +2,15 @@ import { generateUid } from "d2/uid";
 import _ from "lodash";
 import memoize from "nano-memoize";
 import { D2Program } from "../../../types/d2-api";
-import { promiseMap } from "../../../utils/common";
 import { debug } from "../../../utils/debug";
-import { mapCategoryOptionCombo, mapOptionValue, mapProgramDataElement } from "../../../utils/synchronization";
-import { interpolate } from "../../../utils/uid-replacement";
 import { DataValue } from "../../aggregated/entities/DataValue";
 import { AggregatedSyncUseCase } from "../../aggregated/usecases/AggregatedSyncUseCase";
 import { Instance } from "../../instance/entities/Instance";
-import { MetadataMapping, MetadataMappingDictionary } from "../../mapping/entities/MetadataMapping";
-import { CategoryOptionCombo, Program } from "../../metadata/entities/MetadataEntities";
+import { Program } from "../../metadata/entities/MetadataEntities";
 import { SynchronizationResult } from "../../reports/entities/SynchronizationResult";
 import { SynchronizationPayload } from "../../synchronization/entities/SynchronizationPayload";
 import { GenericSyncUseCase } from "../../synchronization/usecases/GenericSyncUseCase";
-import { buildMetadataDictionary, cleanOrgUnitPath } from "../../synchronization/utils";
+import { buildMetadataDictionary } from "../../synchronization/utils";
 import { TEIsPackage } from "../../tracked-entity-instances/entities/TEIsPackage";
 import { TrackedEntityInstance } from "../../tracked-entity-instances/entities/TrackedEntityInstance";
 import {
@@ -23,7 +19,14 @@ import {
 } from "../../tracked-entity-instances/mapper/TEIsPayloadMapperFactory";
 import { EventsPackage } from "../entities/EventsPackage";
 import { ProgramEvent } from "../entities/ProgramEvent";
-import { ProgramEventDataValue } from "../entities/ProgramEventDataValue";
+import { createEventsPayloadMapper } from "../mapper/EventsPayloadMapperFactory";
+// import { promiseMap } from "../../../utils/common";
+// import { mapCategoryOptionCombo, mapOptionValue, mapProgramDataElement } from "../../../utils/synchronization";
+// import { interpolate } from "../../../utils/uid-replacement";
+// import { MetadataMapping, MetadataMappingDictionary } from "../../mapping/entities/MetadataMapping";
+// import { CategoryOptionCombo } from "../../metadata/entities/MetadataEntities";
+// import { cleanOrgUnitPath } from "../../synchronization/utils";
+// import { ProgramEventDataValue } from "../entities/ProgramEventDataValue";
 
 export class EventsSyncUseCase extends GenericSyncUseCase {
     public readonly type = "events";
@@ -184,184 +187,195 @@ export class EventsSyncUseCase extends GenericSyncUseCase {
             .value();
     }
 
-    public async mapPayload(instance: Instance, { events: oldEvents }: EventsPackage): Promise<SynchronizationPayload> {
+    public async mapPayload(instance: Instance, payload: EventsPackage): Promise<SynchronizationPayload> {
         const metadataRepository = await this.getMetadataRepository();
         const remoteMetadataRepository = await this.getMetadataRepository(instance);
-
-        const originCategoryOptionCombos = await metadataRepository.getCategoryOptionCombos();
-        const destinationCategoryOptionCombos = await remoteMetadataRepository.getCategoryOptionCombos();
-        const defaultCategoryOptionCombos = await metadataRepository.getDefaultIds("categoryOptionCombos");
-
         const mapping = await this.getMapping(instance);
-        const events = (
-            await promiseMap(oldEvents, dataValue =>
-                this.buildMappedDataValue(
-                    instance,
-                    dataValue,
-                    mapping,
-                    originCategoryOptionCombos,
-                    destinationCategoryOptionCombos,
-                    defaultCategoryOptionCombos[0]
-                )
-            )
-        ).filter(this.isDisabledEvent);
 
-        return { events };
+        const eventMapper = createEventsPayloadMapper(metadataRepository, remoteMetadataRepository, mapping);
+        return (await eventMapper).map(payload);
     }
 
-    private async buildMappedDataValue(
-        instance: Instance,
-        { orgUnit, program, programStage, dataValues, attributeOptionCombo, ...rest }: ProgramEvent,
-        globalMapping: MetadataMappingDictionary,
-        originCategoryOptionCombos: Partial<CategoryOptionCombo>[],
-        destinationCategoryOptionCombos: Partial<CategoryOptionCombo>[],
-        defaultCategoryOptionCombo: string
-    ): Promise<ProgramEvent> {
-        const { organisationUnits = {} } = globalMapping;
+    // public async mapPayload(instance: Instance, payload: EventsPackage): Promise<SynchronizationPayload> {
+    //     const metadataRepository = await this.getMetadataRepository();
+    //     const remoteMetadataRepository = await this.getMetadataRepository(instance);
+    //     const mapping = await this.getMapping(instance);
 
-        const {
-            mappedProgram,
-            programStages,
-            innerMapping,
-            overlaps = {},
-        } = await this.getRelatedProgramMappings(instance, globalMapping, program, programStage);
+    //     const originCategoryOptionCombos = await metadataRepository.getCategoryOptionCombos();
+    //     const destinationCategoryOptionCombos = await remoteMetadataRepository.getCategoryOptionCombos();
+    //     const defaultCategoryOptionCombos = await metadataRepository.getDefaultIds("categoryOptionCombos");
 
-        const mappedProgramStage =
-            this.getProgramStageMapping(program, programStage, programStages).mappedId ?? programStage;
+    //     const { events: oldEvents } = payload;
 
-        const mappedOrgUnit = organisationUnits[orgUnit]?.mappedId ?? orgUnit;
+    //     const events = (
+    //         await promiseMap(oldEvents, dataValue =>
+    //             this.buildMappedDataValue(
+    //                 instance,
+    //                 dataValue,
+    //                 mapping,
+    //                 originCategoryOptionCombos,
+    //                 destinationCategoryOptionCombos,
+    //                 defaultCategoryOptionCombos[0]
+    //             )
+    //         )
+    //     ).filter(this.isDisabledEvent);
 
-        const mappedCategory = mapCategoryOptionCombo(
-            attributeOptionCombo ?? defaultCategoryOptionCombo,
-            [innerMapping, globalMapping],
-            originCategoryOptionCombos,
-            destinationCategoryOptionCombos
-        );
+    //     return { events };
+    // }
 
-        const mappedDataValues = dataValues
-            .map(({ dataElement, value, ...rest }) => {
-                const { mappedId: mappedDataElement = dataElement, mapping: dataElementMapping = {} } =
-                    mapProgramDataElement(program, programStage, dataElement, globalMapping);
+    // private async buildMappedDataValue(
+    //     instance: Instance,
+    //     { orgUnit, program, programStage, dataValues, attributeOptionCombo, ...rest }: ProgramEvent,
+    //     globalMapping: MetadataMappingDictionary,
+    //     originCategoryOptionCombos: Partial<CategoryOptionCombo>[],
+    //     destinationCategoryOptionCombos: Partial<CategoryOptionCombo>[],
+    //     defaultCategoryOptionCombo: string
+    // ): Promise<ProgramEvent> {
+    //     const { organisationUnits = {} } = globalMapping;
 
-                const mappedValue = mapOptionValue(value, [dataElementMapping, globalMapping]);
+    //     const {
+    //         mappedProgram,
+    //         programStages,
+    //         innerMapping,
+    //         overlaps = {},
+    //     } = await this.getRelatedProgramMappings(instance, globalMapping, program, programStage);
 
-                return {
-                    originalDataElement: dataElement,
-                    dataElement: mappedDataElement,
-                    value: mappedValue,
-                    ...rest,
-                };
-            })
-            .filter(this.isDisabledEvent);
+    //     const mappedProgramStage =
+    //         this.getProgramStageMapping(program, programStage, programStages).mappedId ?? programStage;
 
-        const overlappedDataValues = _(mappedDataValues)
-            .groupBy(item => item.dataElement)
-            .mapValues(items => {
-                const defaultItem = items[0];
-                const { replacer } = overlaps[defaultItem.dataElement] ?? {};
-                if (!replacer) return defaultItem;
+    //     const mappedOrgUnit = organisationUnits[orgUnit]?.mappedId ?? orgUnit;
 
-                const dictionary = _.fromPairs(items.map(item => [item.originalDataElement, item.value]));
-                const value = interpolate(replacer, dictionary);
+    //     const mappedCategory = mapCategoryOptionCombo(
+    //         attributeOptionCombo ?? defaultCategoryOptionCombo,
+    //         [innerMapping, globalMapping],
+    //         originCategoryOptionCombos,
+    //         destinationCategoryOptionCombos
+    //     );
 
-                return _.omit({ ...defaultItem, value }, ["originalDataElement"]);
-            })
-            .values()
-            .value();
+    //     const mappedDataValues = dataValues
+    //         .map(({ dataElement, value, ...rest }) => {
+    //             const { mappedId: mappedDataElement = dataElement, mapping: dataElementMapping = {} } =
+    //                 mapProgramDataElement(program, programStage, dataElement, globalMapping);
 
-        return _.omit(
-            {
-                orgUnit: cleanOrgUnitPath(mappedOrgUnit),
-                program: mappedProgram,
-                programStage: mappedProgramStage,
-                attributeOptionCombo: mappedCategory,
-                dataValues: overlappedDataValues,
-                ...rest,
-            },
-            ["orgUnitName", "attributeCategoryOptions"]
-        );
-    }
+    //             const mappedValue = mapOptionValue(value, [dataElementMapping, globalMapping]);
 
-    private async getRelatedProgramMappings(
-        instance: Instance,
-        globalMapping: MetadataMappingDictionary,
-        originProgram: string,
-        originProgramStage: string
-    ) {
-        const { eventPrograms = {}, trackerPrograms = {}, trackerProgramStages = {} } = globalMapping;
+    //             return {
+    //                 originalDataElement: dataElement,
+    //                 dataElement: mappedDataElement,
+    //                 value: mappedValue,
+    //                 ...rest,
+    //             };
+    //         })
+    //         .filter(this.isDisabledEvent);
 
-        const complexId = `${originProgram}-${originProgramStage}`;
+    //     const overlappedDataValues = _(mappedDataValues)
+    //         .groupBy(item => item.dataElement)
+    //         .mapValues(items => {
+    //             const defaultItem = items[0];
+    //             const { replacer } = overlaps[defaultItem.dataElement] ?? {};
+    //             if (!replacer) return defaultItem;
 
-        if (eventPrograms[originProgram]) {
-            const {
-                mappedId: mappedProgram = originProgram,
-                mapping: innerMapping = {},
-                overlaps,
-            } = eventPrograms[originProgram] ?? {};
+    //             const dictionary = _.fromPairs(items.map(item => [item.originalDataElement, item.value]));
+    //             const value = interpolate(replacer, dictionary);
 
-            const { programStages = {} } = innerMapping;
+    //             return _.omit({ ...defaultItem, value }, ["originalDataElement"]);
+    //         })
+    //         .values()
+    //         .value();
 
-            return { mappedProgram, innerMapping, programStages, overlaps };
-        } else if (trackerPrograms[originProgram]) {
-            const { mappedId: mappedProgram = originProgram, mapping: innerMapping = {} } =
-                trackerPrograms[originProgram] ?? {};
+    //     return _.omit(
+    //         {
+    //             orgUnit: cleanOrgUnitPath(mappedOrgUnit),
+    //             program: mappedProgram,
+    //             programStage: mappedProgramStage,
+    //             attributeOptionCombo: mappedCategory,
+    //             dataValues: overlappedDataValues,
+    //             ...rest,
+    //         },
+    //         ["orgUnitName", "attributeCategoryOptions"]
+    //     );
+    // }
 
-            return { mappedProgram, innerMapping, programStages: trackerProgramStages };
-        } else if (trackerProgramStages[complexId]) {
-            const destinationProgramStage = trackerProgramStages[complexId].mappedId;
+    // private async getRelatedProgramMappings(
+    //     instance: Instance,
+    //     globalMapping: MetadataMappingDictionary,
+    //     originProgram: string,
+    //     originProgramStage: string
+    // ) {
+    //     const { eventPrograms = {}, trackerPrograms = {}, trackerProgramStages = {} } = globalMapping;
 
-            const mappedProgram =
-                (await this.getMappedProgramByProgramStage(instance, destinationProgramStage)) ?? originProgram;
+    //     const complexId = `${originProgram}-${originProgramStage}`;
 
-            return {
-                mappedProgram,
-                innerMapping: {},
-                programStages: trackerProgramStages,
-            };
-        } else {
-            return {
-                mappedProgram: originProgram,
-                innerMapping: {},
-                programStages: trackerProgramStages,
-            };
-        }
-    }
+    //     if (eventPrograms[originProgram]) {
+    //         const {
+    //             mappedId: mappedProgram = originProgram,
+    //             mapping: innerMapping = {},
+    //             overlaps,
+    //         } = eventPrograms[originProgram] ?? {};
 
-    private async getMappedProgramByProgramStage(
-        instance: Instance,
-        destinationProgramStage?: string
-    ): Promise<string | undefined> {
-        if (destinationProgramStage && destinationProgramStage !== "DISABLED") {
-            const remoteMetadataRepository = await this.getMetadataRepository(instance);
+    //         const { programStages = {} } = innerMapping;
 
-            const result = await remoteMetadataRepository.getMetadataByIds<{
-                id: string;
-                program: { id: string };
-            }>([destinationProgramStage], "id, program");
+    //         return { mappedProgram, innerMapping, programStages, overlaps };
+    //     } else if (trackerPrograms[originProgram]) {
+    //         const { mappedId: mappedProgram = originProgram, mapping: innerMapping = {} } =
+    //             trackerPrograms[originProgram] ?? {};
 
-            return result.programStages ? result.programStages[0].program.id : undefined;
-        } else {
-            return "DISABLED";
-        }
-    }
+    //         return { mappedProgram, innerMapping, programStages: trackerProgramStages };
+    //     } else if (trackerProgramStages[complexId]) {
+    //         const destinationProgramStage = trackerProgramStages[complexId].mappedId;
 
-    private getProgramStageMapping = (
-        originProgram: string,
-        originProgramStage: string,
-        programStagesMapping: Record<string, MetadataMapping>
-    ): MetadataMapping => {
-        const complexId = `${originProgram}-${originProgramStage}`;
-        const candidate = programStagesMapping[complexId]?.mappedId
-            ? programStagesMapping[complexId]
-            : programStagesMapping[originProgramStage];
+    //         const mappedProgram =
+    //             (await this.getMappedProgramByProgramStage(instance, destinationProgramStage)) ?? originProgram;
 
-        return candidate ?? {};
-    };
+    //         return {
+    //             mappedProgram,
+    //             innerMapping: {},
+    //             programStages: trackerProgramStages,
+    //         };
+    //     } else {
+    //         return {
+    //             mappedProgram: originProgram,
+    //             innerMapping: {},
+    //             programStages: trackerProgramStages,
+    //         };
+    //     }
+    // }
 
-    private isDisabledEvent(event: ProgramEvent | ProgramEventDataValue): boolean {
-        return !_(event)
-            .pick(["orgUnit", "attributeOptionCombo", "dataElement", "value"])
-            .values()
-            .includes("DISABLED");
-    }
+    // private async getMappedProgramByProgramStage(
+    //     instance: Instance,
+    //     destinationProgramStage?: string
+    // ): Promise<string | undefined> {
+    //     if (destinationProgramStage && destinationProgramStage !== "DISABLED") {
+    //         const remoteMetadataRepository = await this.getMetadataRepository(instance);
+
+    //         const result = await remoteMetadataRepository.getMetadataByIds<{
+    //             id: string;
+    //             program: { id: string };
+    //         }>([destinationProgramStage], "id, program");
+
+    //         return result.programStages ? result.programStages[0].program.id : undefined;
+    //     } else {
+    //         return "DISABLED";
+    //     }
+    // }
+
+    // private getProgramStageMapping = (
+    //     originProgram: string,
+    //     originProgramStage: string,
+    //     programStagesMapping: Record<string, MetadataMapping>
+    // ): MetadataMapping => {
+    //     const complexId = `${originProgram}-${originProgramStage}`;
+    //     const candidate = programStagesMapping[complexId]?.mappedId
+    //         ? programStagesMapping[complexId]
+    //         : programStagesMapping[originProgramStage];
+
+    //     return candidate ?? {};
+    // };
+
+    // private isDisabledEvent(event: ProgramEvent | ProgramEventDataValue): boolean {
+    //     return !_(event)
+    //         .pick(["orgUnit", "attributeOptionCombo", "dataElement", "value"])
+    //         .values()
+    //         .includes("DISABLED");
+    // }
 }
